Refactor with Clean Architeture
