--- Models/Enums/Nationality.cs
@@ -1,20 +1,0 @@
-﻿namespace XP.Formula1.Models.Enums
-{
-    public enum Nationality
-    {
-        NED,
-        MEX,
-        GBR,
-        ESP,
-        MON,
-        CAN,
-        FRA,
-        AUS,
-        THA,
-        GER,
-        FIN,
-        BRA,
-        USA,
-        CHN
-    }
-}
