+++ XP.Formula1.API/Properties/launchSettings.json
