+++ XP.Formula1.Business/Services/DriverService/DriverService.cs
@@ -1,5 +1,7 @@
-﻿using XP.Formula1.Models;
-using XP.Formula1.Models.Enums;
+﻿using XP.Formula1.Domain.Enums;
+using XP.Formula1.Domain.Interfaces.Services;
+using XP.Formula1.Domain.Models;
+
 
 namespace XP.Formula1.Services.DriverService
 {
