+++ XP.Formula1.Domain/Enums/Nationality.cs
@@ -1,0 +1,15 @@
+﻿namespace XP.Formula1.Domain.Enums
+{
+    public enum Nationality
+    {
+        BRA,
+        GBR,
+        USA,
+        CHN,
+        NED,
+        FRA,
+        GER,
+        JPN,
+        NOR
+    }
+}
